Change deviation field in RandomizerConfig to float (0-1)

+++ eventum/core/models/time_pattern_config.py
@@ -56,7 +56,7 @@ class MultiplierConfig(BaseModel):
 
 
 class RandomizerConfig(BaseModel):
-    deviation: int
+    deviation: float
     direction: RandomizerDirection
